refactor(epoch-sync): move function find_target_epoch_to_produce_proof_for
